refactor

+++ models/ImageDatasets.py
@@ -65,7 +65,7 @@ class FedDataset(Dataset):
 
     def __getitem__(self, index):
         img = self.data[index]
-        img = self.transformer(img.float())
+        img = self.transformer(img)
         if self.labels is None:
             return img
         else:
